refactor: open-api -> swagger 폴더명 변경
